Refactored the API service mock
