+++ app/src/androidTest/java/com/lightricks/feedexercise/data/FeedRepositoryTest.kt
@@ -8,9 +8,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.google.common.truth.Truth
 import com.lightricks.feedexercise.database.FeedDatabase
-import com.lightricks.feedexercise.network.FeedApiService
-import com.lightricks.feedexercise.network.FeedData
-import io.reactivex.Single
+import com.lightricks.feedexercise.network.MockFeedApiService
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
@@ -18,23 +16,6 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.TimeUnit
-
-class MockFeedApiService : FeedApiService() {
-    override fun fetchStream(): Single<FeedData> = Single.just(feedData)
-
-    companion object {
-        val feedData = FeedData((0..9).map {
-            FeedData.Metadata(
-                configuration = "Config_$it",
-                id = "ID$it",
-                isNew = it and 2 != 0,
-                isPremium = it and 1 != 0,
-                templateName = "Template$it",
-                templateThumbnailUrl = "abc$it"
-            )
-        })
-    }
-}
 
 @RunWith(AndroidJUnit4::class)
 class FeedRepositoryTest {
